Agriculture is the backbone of our country, nearly three-quarters of India’s families depend on farming.
In India most of the families are depend on the Agriculture, but the
profit which farmer gains is very low. While selling his product to the
market with the help of the middleman, the middleman yields more profit
compared to the farmer. The profit which gained by the farmer is not
worth for his hard work. The only solution to avoid this problem is to
drop the middleman and makes the farmer to sell his product directly to
the market. In order to achieve this solution we designed a webpage to
makes the farmer to sell his product directly to the market. In this
webpage farmer will get into his account using his unique id. His
account contains different option like selling, loan, buying
accessories. Here farmer can sell his product directly to the market
dealer with fixed MSP also farmer can increase or decrease bedding
amount under sell option. He can take loan from the bank for his farming
purpose under loan option. He can buy accessories from the accessory
dealer under buy option. The farmer page is very user friendly with
pictorial option. The website also contains market dealer authentication
and accessory dealer authentication. In market dealer the dealer can buy
products from the farmer based on the price offered by the farmer, also
he can sell his product to the others which is categorized based on
crops type. In accessory dealer page the dealer receives the request
from the farmer as number of accessories farmer needs, even though the
dealer can increase or decrease the accessory rate, offer, availability
of the accessories in his page which will display in the farmer page.
With the help of this website the farmer will yields more money also
reduce the work of the farmer